Unify HDF5 and Parquet ls calls into a single function This PR moves the `ls_hdf()` function, which accepts an `is_parquet` argument to a generic `ls_any()` function. This allows the user to call `ak.ls_any(filename)` on either a Parquet or HDF5 file without having to specify its type.
This is done by having a generic command that is sent to the
server and then allowing the server to differentiate between
the two different file types and calling the correct server
function.

While we would like to get here with all of the IO code, this
function seemed like a good starting point since both functions
have the same arguments (just the filename) and it is a
relatively simple function. Discussion on the naming and how we
would like to handle the differentiating of the file types would
be appreciated.